backend/requirements.txt
# API Framework
fastapi>=0.95.0
uvicorn>=0.21.1

# Database
sqlalchemy>=2.0.0
alembic>=1.10.0
psycopg2-binary>=2.9.5
pymongo>=4.3.3

# Authentication & Security
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.2
aiofiles>=23.1.0
jinja2>=3.1.2

# Testing
pytest>=7.3.1
httpx>=0.24.0

# Development
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
